move `split_bash_commands` into a separate util file
