+++ opendevin/runtime/utils/__init__.py
@@ -1,3 +1,4 @@
+from .bash import split_bash_commands
 from .system import find_available_tcp_port
 
-__all__ = ['find_available_tcp_port']
+__all__ = ['find_available_tcp_port', 'split_bash_commands']
